+++ NEWS.md
@@ -1,4 +1,6 @@
 # srvyr 0.3.1.9000
+* Bug fix for calculating multiple quantiles on grouped data
+  (#38, thanks @iantperry)
 
 # srvyr 0.3.1
 * When converting from a survey db-backed survey to a srvyr one